+++ selene-lib/tests/lints/roblox_incorrect_roact_usage/roblox_incorrect_roact_usage.lua
@@ -17,3 +17,14 @@ e(Components.FooComponent, {
 })
 
 call("foo", {})
+
+e("Frame", {
+	Name = "MyFrame",
+	Visible = false,
+	Parent = workspace,
+})
+
+e(BarComponent, {
+	Name = "BarComponent",
+	Parent = "Earth",
+})